Remove the special null Group for Department
This approach was intended to reduce the amount of special casing
required, but on reflection it actually appears to make things more
complicated.

We're still not special-casing any records this way: there can be one or
more departments, but we'll use the first one as the MoJ for starting
the navigation journey.