Make images load faster
Rather than worry about multiple sizes, we're generating a single
"medium" image with a maximum dimension of 512 pixels, at a parsimonious
60% quality. This yields very small files (16 KB) that load faster, but
which still look good. By resizing them in CSS, we end up with images
that look good on high density displays.

To generate the medium versions of existing images, run this from a
Rails console:

    Person.all.each do |person|
      person.image.recreate_versions! if person.image?
    end