Use org browser on person edit page
We now instantiate the org browser dynamically so that we can have
arbitrarily many on a page.

By using a block to render the hidden form input inside the org browser,
the input can easily be found by inspecting the document hierarchy, with
no state explicitly shared between the form and the org browser.

To handle the callback-inserted additional group sections, we manually
invoke the injector on the new content.