Disable cyclic nodes in edit mode
Although validation will prevent cyclic graphs from being created, they
were still shown in the org browser when editing groups.

Now, the current node and its children are disabled (they don't respond
to clicks, and the radio button is in the disabled state) iff we are in
select mode and a group ID has been provided.