Tell browser not to cache responses
This means that users always get a fresh page, and that the browser will
not store potentially sensitive content in its cache. Whether the latter
is a significant concern is not clear, but this has some nice side
effects in that going back to the dashboard will show any updates, and
going back after logging out will show an unauthenticated page.

Based on
http://arjunghosh.wordpress.com/2008/04/29/how-to-force-the-browser-to-not-cache-in-rails/