Don't show the h1 if we have nothing to say
We no longer show 'SCS 360° Appraisals' as an h1 where we have nothing
better to add - this is shown in the page header in any case.

I have a nagging feeling that we ought to be able to clean up and
simplify the navigation and headers, because it's feeling increasingly
special-cased and hacky.