Put an asterisk after mandatory fields
Unfortunately, there's no cleaner way to do this than to subclass
FormBuilder, as form_for is responsible for instantiating dependencies.
This also makes it tricky to test, so I haven't. It's exercised by the
suite, so we know that it doesn't break anything, and the rest of it is
visual.