Send emails in response to suggestions
In order to achieve this, we must put a hash on the queue instead of the
Suggestion object. Why? Well, see [1] if you really want to know, but
the short answer is that ActiveJob is a mess, and can only
(de-/)serialise a short list of blessed classes in a massive case
statement. Now, if the serialiser were a separate object that could be
injected, we might be able to do something about that, but, well, Rails.

That leaves us with two choices: create a database record for every
Suggestion, or enqueue it as a hash and rehydrate it (which is,
thankfully, easy with Virtus). We might eventually want to do the former
anyway, but for now the latter is less drastic.

1. https://github.com/rails/rails/blob/31085a5cd47894b6fa397ca22fe8aa552d339ce8/activejob/lib/active_job/arguments.rb#L51-L71