:cop: Throttle generation of auth tokens
Chances of collisions via token entropy attacks are extremely small just because of how they're generated. Still, in order to discourage this kind of attack and any DDOS by side effect, a limit is now set on how many tokens can be generated within an hourly period per person.

The default limit is 8 and can be increased/decreased by setting the env variable MAX_TOKENS_PER_HOUR on the server.