Add method to link to object via short name
The contract for objects in the breadcrumb trail is now:

* they must have a name
* they may provide a short_name

The link_to_short_name_unless_current method will link to an object via
its short_name if available, or its name if not. If the short_name is
used, the name will be exposed as the title attribute of the link.