'Stop' people downloading photos
This puts in front of all profile images a div of the same width and
height, preventing the most unskilled users from saving the image via a
right click or by dragging it out of the browser.

This is, of course, ultimately futile: even six-year-olds know how to
get around this, and print screen will still work. The CSS class name
chosen reflects this.