Fix unreliable specs
The particular time chosen for these specs happens to be 57 seconds past
the minute, which gives the spec three seconds to run before the minute
matched in the assertion fails.

This is not usually a problem, but Travis CI sometimes runs very slowly
indeed. Changing the start time to the top of the minute gives us a
whole 60 seconds of oxygen.

Freezing time for the duration of the spec would be a simpler solution,
but, alas, "Capybara does not work with libraries which freeze time".