Fix peer authentication in local development
Now that the healthcheck can cope with both URL and database/host-based
configuration, we can revert to using database configuration when
running locally. This means that PostgreSQL peer authentication will
work again, which means that I can run the tests.

If DB_TEST_URL is defined, this will be used in preference.