Stop generating errors on invalid search input
When percent-encoded invalid UTF-8 was passed via the query parameter,
an exception was raised, and a 500 status resulted.

This happened because, although the invalid UTF-8 was safe at rest, as
soon as it was used to generate the JSON to be sent to Elasticsearch, an
encoding error was raised. This masked a secondary problem where the
parameter was interpolated into the view, which would also have caused
an encoding error.

Instead, we now validate the encoding in the controller (by attempting
to convert the UTF-8 to UTF-32LE) and replace it with an empty string if
it's not valid. The valid query or empty string is assigned to an
instance variable for use in the view.