A more helpful login prompt
After the user enters their email address, a login token is generated and sent
to the user via email. The next page has a bit of informative text prompting the
user to check their emails for a login token.

This page used to look like an error message or a broken page rather than an
important piece of information. We've styled the page and changed the copy to
make it more appealing and hopefully it will be ignored less.