Make legacy profile images continue to work
Profile photos are now stored against a separate object, but the path is
dynamically generated by CarrierWave based on class and ID. It's not
trivial to move all the objects in S3, nor to change CarrierWave to a
more (how shall we say?) forward-thinking path strategy.

Instead, we'll continue to mount the image uploader on the Person as
before, return it for profile_image only if there is not an associated
profile photo object.