Use Bombshelter to prevent image size DoS attack
https://github.com/DarthSim/carrierwave-bombshelter was one of the
CarrierWave recommended solutions to the problem of image bombs. After a
quick review of it and the supporting FastImage gem, I decided that
using it was more sensible than trying to roll the same thing from
scratch.

This commit also include the fast-failing rspec formatter Fuubar and an
update to rspec.