Fix vulnerability to cryptographic timing attacks
Ideally, this would have been implemented using Devise.secure_compare.
However, peoplefinder does not use Devise, so I decided to port the
logic and tests of that method into the app directly.  I also added a
controller spec for the the tokens_controller to ensure that it was
working end-to-end.

I am not entirely happy with my use of Token.find_each to locate the
correct token.  That said, it appeared to be cleanest way to implement
the lookup from a conceptual standpoint. I considered sorting the tokens
and then starting the find_each at the appropriate point in the recored
set. However, doing so seemed like premature optimisation at this
point.