Prevent info leakage about invalid domains
The original pentest flagged a possible vulnerability regarding an
attacker brute-forcing the token authentication to ennumerate all of the
valid emails in the department.  It turns out this is not possible as
the token generator only checks for valid emails and valid domains.  It
is also rate-limited to 8 requests per hour. Nevertheless, after
discussions with our pentesters, I have disabled warnings regarding
invalid domains.