Update the spec stack and clean up config
I believe this app was originally commissioned when it was still common
practice to have everything in spec_helper.  A rails_helper appears to
have been added later, but not everything had been ported over.  This
fixes that and brings the entire stack up-to-date. I've also added
database_cleaner as transactional_fixtures were being a bit too
enthusiastic with the latest version of the capybara drivers.