Remove expired tokens when new tokens are created
There wasn't any mechanism for removing old tokens.  This takes care of
that.  Spent tokens expire after three hours and are then deleted the
next time a new token is created.  It also raises an exception if the
TTL is reduced to less than hour as this would create a race condition
on the rate limiter.