Return existing login token if it is active
When login token is requested, return existing login token if
it is still active, instead of returning a new token and
invalidating the previously sent token.

In production users are sometimes experiencing delays in receiving
login emails, and they request the login email more than once.
If the first token is invalidated when a second is requested, then
the token link in the first email no longer works, which causes
frustration for users.