upgrade rails to close security vulnerabilites
*Unsafe Query Generation Risk in Active Record*
There is a vulnerability when Active Record is used in conjunction with JSON
parameter parsing. This vulnerability has been assigned the CVE identifier
CVE-2016-6317.  This vulnerability is similar to CVE-2012-2660, CVE-2012-2694
and CVE-2013-0155.

*Possible XSS Vulnerability in Action View*
Text declared as "HTML safe" will not have quotes escaped when used as attribute values in tag
helpers.  This vulnerability has been assigned the CVE identifier
CVE-2016-6316.