StateManager
The creation and editing page is now split into three actions - intial create,
editing the photo, editing other information, but this appears to the user as
one editing session. Before this PR, we were getting confusing flash messages
at each stage, as well as an email generated at each stage.

This is now controlled by a cookie (abstracted into s StateManagerCookie object)
which records whether or not the session started with a create or an update, and
whether this action is the result of the final save button being pressed.  This enables
use to send the correct email and display the correct flash message only upon
completions