register javascript driver to redirect console logging to /dev/null
The govuk_frontend_toolkit-5.2.0 has a deprecation warning
in selection-button.js that cannot be turned off.

Once this is resolved in the gem or in the app then we can revert
to using standard poltergiest driver. If you want to enble logging
for dev work just uncomment the `Capybara.javascript_driver = :poltergeist`
line in `rails_helper.rb`.