add unsupported_browser page
users going to the home page in an old browser are to be redirected
to warning page with an option to continue.

If they choose to continue there choice is remembered for the
duration of the session.

The home page will, nonetheless, display a warning to use
Firefox instead regardless.