Photo upload error message fix
Why are we doing this:
Users have reported issues in IE11 with the photo uploader.
In IE11 any error message returned by the server is not visible.

What this PR does:
In ie11, when a file inout is cleared, it will also fire the change event. This caused the error message to be hidden right after it was shown

@jsugarman smashed it and found what the issue was