Support mock Omniauth strategy in development
So that the People Finder can be worked on in development without
integration with the DIT SSO service, this adds the Omniauth developer
strategy.

The strategy is only configured when the `DEVELOPER_AUTH_STRATEGY`
environment variable is present. When it is not set, the app defaults to
the standard DIT SSO behaviour.

To support this change, I have whitelisted CSRF protection for the
`sessions#create` action, to support the Omniauth callback.