Remove requirement to click 'Log in' for JS tests
Fixing the test SSO link in e776c48 means that tests which
previously relied on this method to log in will now cause an
error, as the 'Log in' link will never be found (as the first
log in will actually work).

This method is widely called in the spec codebase and we will
need to remove this in the future.