Merge duplicate accounts when logging in with SSO
Scenario:
- Person logs in via the SSO when they 
have an existing account (created manually or
by logging in with another provider)

- The login script tries to create a new record
but the email exists, so the page blows up.

Solution:
- If a user can’t be found with an internal_auth_key
that matches the email returned by the SSO provider,
we try to do a match on the person’s email.

- If we find a record for that person, we update their
internal_auth_key so that their account gets associated
with their most recent login mechanism

Note:
- Tests could be DRY-er, but want to be explicit 
that there *might* be a difference between an existing
account that was created manually and one created by
logging in with another provider