Fix issue with dir access in spec
Issue was due to a previous upgrade change where Rails.root.join was
used rather than string interpolation. Meant that a Pathname obj was
returned rather than a string literal to represent the path i.e.
"this/is/a/path"

Note: the commit before this used build_stubbed to avoid the issue, I
have left the change in as it passes and build_stubbed is faster in any
case. But this commit addresses the root cause

Error here:
-----------

  1) PeopleHelper profile_image_tag adds alternative text to the person profile by default
     Failure/Error: let(:person)  { create(:person, :with_photo) }

     Errno::EACCES:
       Permission denied @ dir_s_mkdir - /spec
     # ./spec/helpers/people_helper_spec.rb:39:in `block (3 levels) in <top (required)>'
     # ./spec/helpers/people_helper_spec.rb:51:in `block (3 levels) in <top (required)>'