change be_all_there to be_visible
be_all_there does not work with the .form on this page, I think this is
due to some site prisim changes interacting differently with this verion
of rails.

Have visually inspected the page on the pre-upgrade app running the same
test and save_and_open_page and it appears to be identical.

Therefore have changed the check here to be_visible which also returns
true on the pre-upgrate app.

Tests now pass.