Fix spec to check for nil versions in audit trail
Seems that touch events create nil versions - added config as per:

https://stackoverflow.com/questions/54788701/paper-trail-gem-saving-versions-with-object-changes-nil

So that touch events are ignored by default. Which I think was the case
in earlier versions given that the CSS selectors on in the
audit_trail_spec where ordinal values off in the specs because the app
was recording touch events as well as things we are interested in.