Add opt out for Brakeman warning
I think this is safe to ignore will add to ignore file.

There would be a problem if this was a direct request of a
file on the server with a GET like /some/file

Whereas with our app, a 'name' value that is sent to a rails
.find_by(name: name) which selects a report if it's present
in the DB. It is then converted to a csv and returned to the
client.

Given this, if someone was to request something other than
the report name it would just not return from the .find_by
and that would be it. It would not allow someone to access
files on the server.