Add warning to brakeman ignore
It alerts as it there is string interpolation into some SQL that is used
in a where clause. The interpolated string is encased in a private
method and is a static string. Therefore no outside information is added
to the SQL.

If this interplation is replaced by a static dummy string then the
warning goes away therefore I believe this to be a false positive, so it
is added to the ignore file.