Add opt out to Brakeman
Uses interpolation into an SQL string using self.class, but in this case it is literally just the name of the enclosing class - which is always the same due to the scope of the method.

self.class here returns "UpdateGroupMembersCompletionScoreJob" which is essentially a constant. This means that the injected string isn't based on user input and is static. Therefore there is no real issue here.

Added to opt out