CT-4200 Remove report a problem from logged out pages
Due to the spamming that has been received into Google Groups, it has been suggested that we remove the ‘Report a problem’ located in the footer from all the logged out pages.

In addition, so people still have a way to report login issues, the support email address has been added to the sign in / request link page.

The copy/placement can be discussed.

Note: as part of this work, I had some issues setting up ElasticSearch locally and making it work, so I tweaked the README instructions with what, in the end, helped me getting this up and running.

Also, when running the app/tests there were several deprecation warnings that I resolved (Pundit, Selenium...)