Update CircleCI badge and remove Gemnasium
In a recent audit, it was raised as a false alarm the fact the CircleCI badge contains a token.

However this token scope is only "status" and can only be used to retrieve the status of a build (pass/fail).

The token is there because this repo used to be private, and private repos badges can only work with a token.

As the repo is no longer private, and in order to no raise more false alarms in the future, I'm updating the badge to the "public" one, no longer needing a token, and will revoke the token so it cannot be used.

Also, removed the badge for gemnasium because that service is no longer active.